Squashed commit of the following:

commit 47d5796c955a80d78ad938e8962ff98f412a4889
Author: Michal Pawlus <michal.pawlus@relativity.com>
Date:   Fri Jul 27 10:54:18 2018 +0200

    Constant for query length.

commit faa4157ede6c4552821caa33770f85be080a13ce
Merge: 0d48c40 5c9f4a8
Author: Piotr Szmigielski <piotr.szmigielski@relativity.com>
Date:   Thu Jul 26 16:40:55 2018 +0200

    Merge branch 'develop' into RegressionTestsFixes

commit 0d48c408ac8e3ba4936799b9883fb9b61d290b8b
Author: Michal Pawlus <michal.pawlus@relativity.com>
Date:   Thu Jul 26 15:43:26 2018 +0200

    Fixes.

commit 8c4d1d3ad9f2e36af5362b3f1013c9c038636aaf
Author: Michal Pawlus <michal.pawlus@relativity.com>
Date:   Thu Jul 26 13:45:41 2018 +0200

    Serilog libs downgraded to stable versions.

commit 036c5f373cf5d214fa52cc0435fe81aacea9dcab
Author: Michal Pawlus <michal.pawlus@relativity.com>
Date:   Thu Jul 26 13:30:46 2018 +0200

    Methods merged in Page.

commit e89e167894199ee77afaf0fc34afe2ada4f4a6bf
Author: Michal Pawlus <michal.pawlus@relativity.com>
Date:   Thu Jul 26 13:22:23 2018 +0200

    Workaround for red Server Error message removed.

commit ec77d74e1a871d18520e40fd476d1ee27bfe29fd
Author: Michal Pawlus <michal.pawlus@relativity.com>
Date:   Thu Jul 26 12:53:38 2018 +0200

    String literal moved to constant.

commit 57c8a0a385a62b9f3a4435f0adcd2d18047c6e90
Author: Michal Pawlus <michal.pawlus@relativity.com>
Date:   Thu Jul 26 12:43:23 2018 +0200

    Removing test workspaces moved to console app runner.

commit 19a32e444181b3c68aa424264f827bdeb4855560
Author: Piotr Szmigielski <piotr.szmigielski@relativity.com>
Date:   Mon Jul 16 17:27:14 2018 +0200

    Added workspace template name to app.config

commit a2b349472b6093c8e153e5e55526a778d5e453cd
Author: Piotr Szmigielski <piotr.szmigielski@relativity.com>
Date:   Mon Jul 16 15:20:18 2018 +0200

    Uncommented headless argument in ChromeOptions to allow running UI tests on Jenkins

commit c5c1a246c88f52d6e2d6b77fa31611001feab6f2
Merge: d8c97c8 d32c27f
Author: Piotr Szmigielski <piotr.szmigielski@relativity.com>
Date:   Mon Jul 16 14:44:22 2018 +0200

    Merge branch 'develop' into RegressionTestsFixes

commit d8c97c8444443603b845aed721282c86b9d458e8
Author: Michal Pawlus <michal.pawlus@relativity.com>
Date:   Thu Jul 12 14:30:12 2018 +0200

    Regression Tests fixes

+++ Source/kCura.IntegrationPoints.UITests/Tests/UiTest.cs
@@ -30,27 +30,25 @@ namespace kCura.IntegrationPoints.UITests.Tests
 
 	public abstract class UiTest
 	{
-		public static string Now => DateTime.Now.ToString("yyyy-MM-dd_HH-mm-ss");
 
-		protected IWindsorContainer Container;
-		protected static readonly ILogger Log = LoggerFactory.CreateLogger(typeof(UiTest));
+		private readonly Lazy<ITestHelper> _help;
+
 		protected IConfigurationStore ConfigurationStore;
 
-		public ITestHelper Helper => _help.Value;
-		private readonly Lazy<ITestHelper> _help;
+		protected IWindsorContainer Container;
+
+		protected static readonly ILogger Log = LoggerFactory.CreateLogger(typeof(UiTest));
 
 		protected static readonly List<Tuple<string, string>> DefaultFieldsMapping = new List<Tuple<string, string>>
 		{
 			new Tuple<string, string>("Control Number", "Control Number"),
 			new Tuple<string, string>("Extracted Text", "Extracted Text"),
-			new Tuple<string, string>("Title", "Title"),
-			new Tuple<string, string>("Date Created", "Date Created")
+			new Tuple<string, string>("Title", "Title")
 		};
 
-		protected static readonly List<Tuple<string, string>> ControlNumberFieldsMapping = new List<Tuple<string, string>>
-		{
-			new Tuple<string, string>("Control Number", "Control Number"),
-		};
+		public static string Now => DateTime.Now.ToString("yyyy-MM-dd_HH-mm-ss");
+
+		public ITestHelper Helper => _help.Value;
 
 		protected TestConfiguration Configuration { get; set; }
 
@@ -179,18 +177,47 @@ namespace kCura.IntegrationPoints.UITests.Tests
 		[OneTimeTearDown]
 		protected void CloseAndQuitDriver()
 		{
-			if (!NUnit.Framework.TestContext.CurrentContext.Result.Outcome.Equals(ResultState.Success))
+			try
 			{
-				SaveScreenshot();
+				if (!NUnit.Framework.TestContext.CurrentContext.Result.Outcome.Equals(ResultState.Success))
+				{
+					SaveScreenshot();
+				}
+			}
+			catch (Exception ex)
+			{
+				Log.Error(ex, "Error during saving screenshot.");
 			}
 
-			if (string.IsNullOrEmpty(SharedVariables.UiUseThisExistingWorkspace))
+			try
 			{
-				Workspace.DeleteWorkspace(Context.GetWorkspaceId());
+				if (string.IsNullOrEmpty(SharedVariables.UiUseThisExistingWorkspace) && Context.WorkspaceId != null)
+				{
+					Workspace.DeleteWorkspace(Context.GetWorkspaceId());
+				}
 			}
-			Context.TearDown();
+			catch (Exception ex)
+			{
+				Log.Error(ex, "Error during deleting workspace.");
+			}
 
-			Driver?.Quit();
+			try
+			{
+				Context.TearDown();
+			}
+			catch (Exception ex)
+			{
+				Log.Error(ex, "Error in Context TearDown.");
+			}
+
+			try
+			{
+				Driver?.Quit();
+			}
+			catch (Exception ex)
+			{
+				Log.Error(ex, "Quiting Driver failed.");
+			}
 		}
 
 		protected GeneralPage EnsureGeneralPageIsOpened()
@@ -200,7 +227,7 @@ namespace kCura.IntegrationPoints.UITests.Tests
 			{
 				return loginPage.Login(SharedVariables.RelativityUserName, SharedVariables.RelativityPassword);
 			}
-			return new GeneralPage(Driver);
+			return new GeneralPage(Driver).PassWelcomeScreen();
 		}
 
 		protected void SaveScreenshot()
